requirements.txt
numpy=1.17.0
scipy=1.2.1
matplotlib==3.0.0
torch=1.2.0
torchvision==0.4.0
Pillow==5.4.1


